% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mecab_test}
\alias{mecab_test}
\title{Calcuate Frequency table
This simple function calcuate Frequency table}
\usage{
mecab_test(str)
}
\arguments{
\item{str}{character}
}
\value{
Data Frame
}
\description{
Calcuate Frequency table
This simple function calcuate Frequency table
}
\keyword{MeCab}

